Completed Task1
In the main method of the WMethod class I converted the vector
containing the test cases to a string array, then converted that array
to a string. I then converted that string to a character array so I
could convert it back to a string with spaces between the values within
the array so the runFSM method from the Utilities class could run
properly. Stored the output in the Task1 folder.